src/frontend/api/requirements.txt
# Minimal requirements for Vercel Python function
# Only include essential dependencies to stay under 250MB limit

# Core dependencies
pydantic
python-dotenv

# Supabase and database
supabase
psycopg2-binary
pgvector

# Google AI
google-generativeai
langchain-google-genai

# Vector operations
numpy

# ASGI handler for Vercel
mangum